fast-mongo_motor_crud/pyproject.toml
[tool.poetry]
name = "fast-mongo-motor-crud"
version = "0.1.0"
description = ""
authors = ["Michael Youngchan CHO <michael.yc.cho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.2"
python-dotenv = "^1.0.1"
motor = "^3.5.1"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
